Merge binary target into root manifest (#3)
Simplify by using the same Package.swift for different targets,
defaulting to production.

Beware that this makes the repository "unusable" for SwiftPM as is (src
is inaccessible), requiring it to be fetched as a submodule first. This
is okay as submodule remains the only way to provide multiple
sub-packages. The folders inside "Plugins" are in fact SwiftPM packages.