OpenVPN: Fix missing network routes to the VPN (#49)
`ifconfig` with topology `subnet` and `ifconfig-ipv6` define the
properties of the VPN network:

- The local VPN address
- The network mask or prefix (`ifconfig-ipv6`) of the VPN

For example, `ifconfig 192.168.5.12 255.255.255.0` means that
192.168.5.12 is our IPv4 address in the 192.168.5.0/24 network. A route
to such a network is missing.

The format of `ifconfig-ipv6`, instead, is similar to `ifconfig` with
topology `net30`, where the second argument is the remote VPN endpoint.
However, the IPv6 address comes with a prefix that implies the IPv6
network of the VPN. Similarly, no such route is added.