Add implementations compatible with System Extensions (#61)
When a packet tunnel provider is deployed as System Extension, some
features are unavailable. The lack of App Groups in particular limits
the scope of app/tunnel IPC. Therefore, alternative implementations are
provided here:

- NETunnelEnvironment, a TunnelEnvironment that fetches the tunnel
environment values at intervals via .sendMessage()
- POSIXDNSStrategy, a SimpleDNSResolver that uses POSIX getaddrinfo() to
perform DNS resolution

CFDNSStrategy (based on CFHost) is not working out of the box, even
though it might be a bug in my implementation. The resolver works in the
appex, yet systematically times out in the sysex.