OpenVPN: Add C rewrite of OpenSSL ObjC crypto (#69)
Tweak conditionals in the package manifest to let the ObjC and C
implementations coexist seamlessly, and also offer ObjC/C bridges to the
higher package tiers. Three levels are implemented:

- .legacy (only ObjC)
- .bridged (ObjC wrappers around C code, reusing the same ObjC headers)
- .native (Swift wrappers around C code, WIP)

Along the way, make initializers failable and reuse the tests for all
the modes.

The .legacy mode is the default, it's stable, and it will stay that way
on Apple platforms for a long while. The .bridged mode was quickly
tested in Passepartout, and the connection was flawless. Yay!

This is big bump for #52 to port OpenVPN on non-Apple, which will be the
testbed for the new C code.