OpenVPN: Add Swift wrappers of DataPath variants (#73)
Unless the code is in legacy mode, interpose a DataPathWrapper between
the OpenVPN logic and the DataPath implementation. This allows unifying
DataPath behavior and tests in a way that is agnostic of its underlying
ObjC and C implementations. It's crucial that the #if blocks **guarantee
that the legacy code remains unaltered**, namely in Negotiator and
DataChannel.

DataPathWrapper has static initializers for:

- .legacy (DataPath, ObjC) = .wrappedDataPath in manifest
- .native (CDataPath, Swift/C) = .wrappedDataPathNative in manifest
(OPENVPN_DP_NATIVE)

with overloads with PRNG, seed, or explicit keys (for testing).

Details:

- Move Swift/C wrappers (CZeroingData, CDataPath) from test targets to
the OpenVPNOpenSSL target
- Replicate Z() macros as CZ() for CZeroingData
- Build the session keys from the auth response in a more understandable
way (DataPathWrapper+Keys)
- Dedup inline methods with clashing names because OpenVPNOpenSSL
includes both ObjC and C for bridging (allocation.h was updated to equal
the C version, xor.h symbols were renamed)
- Use DataPathWrapper in tests (native only)

Bonus:

- Improve the readability of Package.swift by moving details to the
bottom
- Make C crypto construction atomic by performing the configure step in
the initializer if a set of encryption/decryption keys is provided.
Therefore, CDataPath doesn't need to configure encryption/decryption,
the crypto input does it. Keep the interface backward-compatible for
CryptoOpenSSLTests.

This PR, specifically DataPathWrapper/DataPathProtocol, allows for easy
comparison of the performance of the legacy ObjC and native C
implementations (later).