OpenVPN: Port and remap all errors (#85)
Keep granularity to tell C errors from Swift errors.

This is because the old code often treated the errors coming from the
ObjC layer as unrecoverable, so we'd better stick with that behavior.
With the new error structure, though, we streamline the do/try blocks of
TLS put/pull and data encrypt/decrypt to make the flow more readable and
predictable, thanks to the catch by specific error type.

Plus:

- Handle creation failure of dp_mode_t objects (nullable unless mock)
- CryptoErrorPRNG is removed because PlatformPRNG simply triggers
fatalError() if the PRNG fails
- Wrap error as OpenVPNSessionError before dispatching to delegate in
sessionDidStop

WARNING: CryptoError from CryptoOpenSSL is made internal and replicated
in OpenVPNOpenSSL to use the C crypto_error_code (private otherwise).