Reorganize package to externalize vendors (#102)
Rewrite the package manifest to dynamically resolve dependencies based
on the current platform and a list of preferred vendors. The goal is to
minimize or even eliminate the external dependencies when the OS may
provide a native solution.

This PR attacks the OpenSSL dependency in particular, which on *nix
platforms is better relied on as a system library than statically
linked. We keep the dependency on Apple platforms, but this may change
in the future if CryptoKit or other crypto libraries with a smaller
footprint offer comparable features and performance. TLS is a bit more
complicated, but it doesn't need great performance either.

Overall, this is how the package is modified:

- The PartoutPlatforms* packages are dropped and merged into
PartoutVendors, which is now a mix of OS and third-party frameworks.
- On top of the manifest, a list of preferred vendors is specified for
each platform, and each vendor provides a set of features.
- If a package requires a feature, we look up the vendor list, and
inject a package dependency on the first vendor providing that feature.
For example, the .openVPN package requires the .crypto and .tls
features. Apple platforms include the .openSSLApple vendor, so the
.openVPN package will depend on that one. On Linux, the .openSSLShared
vendor is preferred, so the .openVPN package will depend on the
system-wide OpenSSL installation instead.
- This will be even more evident and beneficial in WireGuard, where
non-Apple platforms usually support kernel-level integration.

PartoutCrypto is introduced to provide agnostic Crypto/TLS:

- CryptoOpenSSL_C` is now split into headers (.h) and definitions (.c).
- CryptoCore_C is the bare set of headers.
- CryptoOpenSSL_C, as a concrete target, implements the headers with
OpenSSL.
- Other crypto targets will include CryptoCore_C to implement it with
the corresponding crypto library (e.g. [CNG on
Windows](https://github.com/passepartoutvpn/partout/issues/101)).

To make this separation possible:

- All C structures are moved to .c with a forward pointer declaration in
the .h file.
- Treat crypto_aead/cbc/ctr structures as opaque `crypto_ctx` with "poor
man C inheritance" (downcast to `crypto_ctx` and access "base" field).
All `crypto_*_create` and `crypto_*_free` functions now take a generic
`crypto_ctx`.
- Replace `tls_channel_t *` with the opaque `tls_channel_ctx` as well.
- Function pointers inside C structures are encapsulated in new generic
functions (inline `crypto_*` functions in crypto.h).
- C structs are nicely bridged in Swift as `OpaquePointer` rather than
`Unsafe*Pointer`. No more `.pointee`.

Consequently:

- Rename OpenVPNOpenSSL_C to OpenVPN_C
- Rename OpenVPNOpenSSL_Cross to OpenVPN_Cross

for being now fully agnostic of OpenSSL. With this in place,
OpenVPN_Cross will in fact compile, but not link without a vendored
implementation of CryptoCore_C.

Other than this, the huge majority of the changes in this PR are moved
files and updated imports.

The legacy ObjC implementations of CryptoOpenSSL and OpenVPNOpenSSL are
NOT modified (production).