Fix exported imports and lifecycle of CryptoFlags in tests (#106)
Objects inside the withUnsafeBytes() blocks must not outlive the blocks.
Therefore, execute the crypto tests inside the flags mapper block. Tests
were randomly succeeding (or failing) due to the unpredictable values of
cipher IV, regardless of the flags.

Take the opportunity to clean up dependencies and imports because
Passepartout was failing to build:

- _PartoutVendors* were not auto-exported from Partout
- Partout was not even including _PartoutVendors* as dependencies
- Clean up helpers in CryptoCore and related tests