Windows: Add CNG implementations of CryptoCore_C (#109)
Add code for crypto AEAD, CBC, and CTR on Windows. These make the
`libcrypto` side of the OpenSSL counterpart, whereas the `libssl` side
still needs to be done (tls.c). The keys.c implementation will also be
done in a separate PR.

The test target for CryptoCore_C now depends on the available `.crypto`
vendor dynamically. That is:

- Apple: Tests depend on OpenSSL
- Windows: Tests depend on CNG

Lastly:

- Change encrypt/decrypt signature from (bool + size_t*) to (size_t)
- Receive the length of the output buffer in encrypt/decrypt (required
by BCrypt*)
- Reuse CryptoWrapper for all algorithms
- Move CZeroingData shortcuts to CryptoCore
- Rename length to count in CZeroingData
- Add CZX for hex init

Fixes #101