Extend ProviderModule for authentication and WireGuard (#119)
Extend Registry, APIEngine, and ProviderModule to support:

- Per-device provider module resolution
- Provider authentication, credentials or token based
- WireGuard storage for provider key registration, expect a peer/key
association in the form of sessions

Precisely:

- Add more functions to the API virtual machine (TODO: add prefix or
aggregate under a single entity to distinguish)
- In particular, add getResult() for generic HTTP requests with headers
and body, not only GET
- Return HTTP status codes and generic error strings from JavaScript
responses to improve troubleshooting, error codes alone are not
informative

Eventually, interpose the `CodableOptions` wrapper in
`ProviderModule.moduleOptions` to work around the non-linear
serialization of `ModuleType` when used as a dictionary key. Swift was
in fact encoding `[ModuleType: Data]` to:

```
["WireGuard", "...base64..."]
```

rather than the desired:

```
{"WireGuard": "...base64..."}
```

and this causes trouble when passing the map to JavaScript.

Migrate some related tests to Swift Testing.