OpenVPN: Unhonored local redirect-gateway option (#149)
The default routes were only added to the configuration in the presence
of `ifconfig`, which doesn't apply to local configurations, only remote.

Do not translate the routing policies into routes immediately in the
parser because this is also confusing. Instead, move the computation of
default routes to NetworkSettingsBuilder, right before merging the
local/remote options. If `isIPv4Gateway` is true (or `isIPv6Gateway` for
IPv6), then inject the default route.

Also resolve the route gateway in the builder. Deduplicate
NetworkSettingsBuilder for easier maintenance, logic is the same in
legacy/cross.

Pre-existing bug, only revealed by #133