Build as monolith with CMake (#176)
Build everything together into 2 super-targets:

- Partout_C
- Partout

with dependencies on vendored submodules:

- PartoutCore
- OpenSSL
- mbedTLS
- WireGuardGo (unused until #118)

to generate a single-file dynamic library, as a static library would
require bundling all the dependencies.

The `scripts/build.sh` build generates `libPartout.dylib|so|dll` at the
root of `bin`, with the statically linked dependencies under the same
directory, named after the vendor submodule name. The static libraries
are then used in SwiftPM for local development on non-Apple platforms
via .unsafeFlags()

Building for Android is now done in 2 steps:

- The vendored static libraries are cross-compiled with CMake for
Android
- Partout remains a SwiftPM target with the static dependencies also
embedded via .unsafeFlags()
- The Partout dynlib is eventually built with the Swift Android SDK,
which is not easily available to CMake

There are two ways to coexist with SwiftPM and still deal with the now
invalid `Partout*` imports:

- Replicate the dependency tree of the SwiftPM modules in CMake
- Wrap all imports of `Partout*` in a `#if !PARTOUT_MONOLITH`

We go for the second option as it's easier to do and maintain. Retain
the imports of the C modules, though, for which we define explicit
module maps.

Other changes:

- Rename the old PartoutVendorsTLSCore_C name to just PartoutTLSCore_C
- Split OpenVPN dp_mode_shortcuts from mock functions
- Work around an explicit package prefix in Partout+OpenVPN.swift
- Fix a wrong include in TLSMbedTLS_C/tls.c (dummy code)

Fixes #173