Add TunnelController based on virtual tun devices (#192)
With implementations for macOS and Linux, other *nix should be similar.
The tun device is operated on like a normal file descriptor. This PR
integrates the controller lifecycle and the I/O, but leaves routing for
later.

A little change in the tun lifecycle is that it's now created on
setTunnelSettings(), rather than statically (#191). Therefore,
clearTunnelSettings() is updated to receive the IOInterface that the
tunnel was configured with.