Cross: Support generic C-based tun implementation (#195)
On platforms where the tun device is not accessible with a C API, the
implementation must be delegated externally. This is the case for
Android, where the VpnService API is only accessible from Kotlin/JVM.

With the goal of Kotlin interop, this PR expands on
VirtualTunnelController to accept an optional C struct (unowned). The
ctrl struct contains a generic pointer and a set of callbacks that are
expected to comply with the setTunnelSettings() and
cancelTunnelSettings() functions of TunnelInterface, where the
set_tunnel() callback can return an opaque implementation for
VirtualTunnelInterface as well. This mechanism will allow Android to
perform Kotlin calls to VpnService from C via JNI.

Worth noting that clearTunnelSettings() is never called in
LegacyOpenVPNConnection, and we'll keep it that way because Apple is out
of scope here. Therefore, clearTunnelSettings() is only introduced in
the cross OpenVPNConnection implementation.

Bonus:

- Move library versioning to partout.h/.c because it's where it makes
the most sense
- Use a macro to load the DLL functions of Wintun
- Change the CMake output to dot dirs to avoid polluting Xcode searches
- Delete Swift targets of non-Apple platforms, use C only