OpenVPN/Cross: Temporarily integrate deprecated LZO for feature parity (#208)
LZO must be deleted, but this middle step is necessary to alleviate the
migration from ObjC to the C codebase. The memory management of
compression/decompression is also subpar, yet we don't care, as this
"feature" is only temporary. In the legacy target, the former LZOFactory
dynamic check is replaced with the OPENVPN_DEPRECATED_LZO define.

Extra:

- Group OpenVPN/WireGuard provider extensions with the PARTOUT_API
define
- Fix some C warnings by enabling compiler flags globally
- Use pp_assert consistently
- Merge cross/legacy parser tests