WireGuard/Cross: Handle errors thrown from path handler gracefully (#274)
The backend is stopped on error, and when the backend is stopped, the
tunnel is ineffective and bypassed. This happens because
`endpointUapiConfiguration()` throws in the new implementation, and
`didUpdateReachable()` stops the backend in the catch block (legacy
`didReceivePathUpdate()` doesn't throw).

Simulate failure to see if Personal Hotspot starts failing when that
happens.

Maybe related to
https://github.com/partout-io/passepartout/issues/1327#issuecomment-3517977325