Set boundaries of Foundation compatibility layer (#278)
Introduce PARTOUT_FOUNDATION_COMPAT. Will only be .on if OS != .apple,
now hardcoded to .off

- Use Foundation on Apple, compatible code elsewhere
- Omit all explicit Foundation imports, rely on PartoutCore to re-export
- Expose CodableProfile from Registry, delegate JSON serialization via
TextEncoding
- Move logging constants from JSONEncoder to PartoutLogger (e.g.
`.redactedValue`)
- Merge exports and dictionary JSON extensions into Foundation+Exports
- Rename Registry.compatibleProfile() to .fallbackProfile() to
disambiguate
- Build dynamic library on Linux to test ldd against Foundation
- Exclude ObjC code with comp .on
- Exclude some files until drop-in replacement

The analysis revealed that JSON was only used to (de)serialize profiles
in Registry, so this PR fundamentally decouples the library from
JSONEncoder/JSONDecoder, a major Foundation dependency, through the
TextEncoder/TextDecoder protocols. JSON encoding is only available in
Foundation-native mode (i.e. Apple), and will be implemented differently
on non-Apple platforms.

Everywhere else, JSON is only used for:

- Apple-specific code (NEProtocolCoder implementations)
- Interop between internal classes (Shared/StaticTunnelEnvironment)
- Logging/debugging (Encodable.asJSON, JSON.debugDescription)

And could therefore be limited to those scenarios with `#if
!PARTOUT_FOUNDATION_COMPAT`.

When compatibility is .on, the package (though PartoutFoundation is
still broken) now _compiles_ on non-Apple without any dependency on
Foundation. This was verified with `ldd` on Linux on the generated
`libpartout.so` library.